# src/main/java/mg/federation/federationapi/dto/CollectivityIdentity.java
package mg.federation.federationapi.dto;

public class CollectivityIdentity {

    private String name;
    private String number;

    public String getName() { return name; }
    public void setName(String name) { this.name = name; }

    public String getNumber() { return number; }
    public void setNumber(String number) { this.number = number; }
}