# src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: Agricultural federation API
  version: 0.0.3

servers:
  - url: http://localhost:8080

paths:

  /collectivities:
    post:
      summary: Create collectivities
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CreateCollectivity'
      responses:
        201:
          description: Created

  /collectivities/{id}/informations:
    put:
      summary: Assign name and number
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectivityIdentity'
      responses:
        200:
          description: OK

  /collectivities/{id}/membershipFees:
    post:
      summary: Create membership fees
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CreateMembershipFee'
      responses:
        200:
          description: OK

    get:
      summary: Get membership fees
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK

  /members:
    post:
      summary: Create members
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CreateMember'
      responses:
        201:
          description: OK

  /members/{id}/payments:
    post:
      summary: Create payments
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CreateMemberPayment'
      responses:
        201:
          description: OK

components:
  schemas:

    CreateCollectivity:
      type: object
      properties:
        location:
          type: string
        federationApproval:
          type: boolean

    CollectivityIdentity:
      type: object
      properties:
        name:
          type: string
        number:
          type: string

    CreateMember:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        referees:
          type: array
          items:
            type: string
        registrationFeePaid:
          type: boolean
        membershipDuesPaid:
          type: boolean

    CreateMembershipFee:
      type: object
      properties:
        amount:
          type: number
        label:
          type: string
        frequency:
          type: string

    CreateMemberPayment:
      type: object
      properties:
        amount:
          type: number
        paymentMode:
          type: string


          /collectivities/{id}:
            get:
              summary: Get one collectivity
              parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                    type: string
              responses:
                200:
                  description: OK

          /collectivities/{id}/financialAccounts:
            get:
              summary: Get financial accounts
              parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                    type: string
                - in: query
                  name: at
                  required: true
                  schema:
                    type: string
                    format: date
              responses:
                200:
                  description: OK